feat: 강아지 정보 수정 api 추가

+++ backend/src/dogs/dogs.controller.ts
@@ -1,4 +1,15 @@
-import { Body, Controller, Delete, ForbiddenException, Get, HttpCode, Param, ParseIntPipe, Post } from '@nestjs/common';
+import {
+    Body,
+    Controller,
+    Delete,
+    ForbiddenException,
+    Get,
+    HttpCode,
+    Param,
+    ParseIntPipe,
+    Patch,
+    Post,
+} from '@nestjs/common';
 import { AccessTokenPayload } from 'src/auth/token/token.service';
 import { Serialize } from 'src/common/interceptor/serialize.interceptor';
 import { User } from 'src/users/decorators/user.decorator';
@@ -31,14 +42,30 @@ export class DogsController {
     @Delete('/:id')
     @HttpCode(204)
     async delete(@User() { userId }: AccessTokenPayload, @Param('id') dogId: number) {
+        const owned = await this.usersService.checkDogOwnership(userId, dogId);
+        if (!owned) {
+            throw new ForbiddenException('주인이 아닌 강아지에 대한 요청 ');
+        }
         await this.dogsService.deleteDogFromUser(userId, { id: dogId });
 
         return true;
     }
 
+    @Patch('/:id')
+    @HttpCode(204)
+    async update(@User() { userId }: AccessTokenPayload, @Param('id') dogId: number, @Body() dogDto: DogDto) {
+        const owned = await this.usersService.checkDogOwnership(userId, dogId);
+        if (!owned) {
+            throw new ForbiddenException('주인이 아닌 강아지에 대한 요청 ');
+        }
+        await this.dogsService.updateDog(dogId, dogDto);
+
+        return true;
+    }
+
     @Get('/:id')
-    async getOneProfile(@User() user: AccessTokenPayload, @Param('id', ParseIntPipe) dogId: number) {
-        const owned = await this.usersService.checkDogOwnership(user.userId, dogId);
+    async getOneProfile(@User() { userId }: AccessTokenPayload, @Param('id', ParseIntPipe) dogId: number) {
+        const owned = await this.usersService.checkDogOwnership(userId, dogId);
         if (!owned) {
             throw new ForbiddenException('주인이 아닌 강아지에 대한 요청 ');
         }
@@ -46,14 +73,14 @@ export class DogsController {
     }
 
     @Get('/walk-available')
-    async getAvailableDogs(@User() user: AccessTokenPayload): Promise<DogProfile[]> {
-        const ownDogIds = await this.usersService.getOwnDogsList(user.userId);
+    async getAvailableDogs(@User() { userId }: AccessTokenPayload): Promise<DogProfile[]> {
+        const ownDogIds = await this.usersService.getOwnDogsList(userId);
         return await this.dogsService.getProfileList({ id: In(ownDogIds), isWalking: false });
     }
 
     @Serialize(DogStatisticDto)
     @Get('/statistics')
-    async getDogsStatistics(@User() user: AccessTokenPayload) {
-        return this.dogsService.getDogsStatistics(user.userId);
+    async getDogsStatistics(@User() { userId }: AccessTokenPayload) {
+        return this.dogsService.getDogsStatistics(userId);
     }
 }
